Feat:change to dynamic page
